Restructuring Project/Library
Restructered so all helpers will be in library instead of project. Also
removed direct connection between PushButtonHelper and LEDHelper.

LightSensorController is currently empty work to fix that will be done
at a later time

PushButtonController has a new @Get endpoint which initializes the
PushButtonHelper and adds an event listener to it

LEDHelper has a new switchState method, if the led is on, it will be
turned off, if it is off it will be turned on

PushButtonHelper has a new isPressed field. When initialized an
eventlistener will be added to buttonInput which switches the state of
isPressed according to the status of the button. When pressed, isPressed
will be true, otherwise false. This is a public field which can be
accessed via PushButtonHelper.isPressed
