+++ components/src/main/java/com/opensourcewithslu/components/controllers/LightSensorController.java
@@ -1,7 +1,5 @@
 package com.opensourcewithslu.components.controllers;
 
-import com.opensourcewithslu.components.outputdevices.led.Darkness;
-import com.opensourcewithslu.components.outputdevices.led.LightSensorHelper;
 import io.micronaut.http.HttpResponse;
 import io.micronaut.http.annotation.Controller;
 import io.micronaut.http.annotation.Get;
@@ -12,20 +10,4 @@ import javax.validation.constraints.Positive;
 @Controller("/light")
 public class LightSensorController {
 
-    private final LightSensorHelper sensorHelper;
-
-    public LightSensorController(LightSensorHelper sensorHelper) {
-        this.sensorHelper = sensorHelper;
-    }
-
-    @Get
-    public Darkness status() {
-        return sensorHelper.getDarkness();
-    }
-
-    @Post("/threshold/{i}")
-    public HttpResponse update(@Positive int i) {
-        sensorHelper.setDarknessThreshold(i);
-        return HttpResponse.noContent();
-    }
 }
