+++ micronautpi4j-utils/src/main/java/com/opensourcewithslu/utilities/Pi4JFactory.java
@@ -1,10 +1,10 @@
 package com.opensourcewithslu.utilities;
 
 import com.pi4j.Pi4J;
+import com.pi4j.context.Context;
 import com.pi4j.io.gpio.digital.DigitalInput;
 import com.pi4j.io.gpio.digital.DigitalOutput;
 import io.micronaut.context.annotation.Bean;
-import io.micronaut.context.annotation.Context;
 import io.micronaut.context.annotation.EachBean;
 import io.micronaut.context.annotation.Factory;
 import jakarta.inject.Singleton;
@@ -18,8 +18,7 @@ public class Pi4JFactory {
     }
 
     @EachBean(DigitalOutputConfiguration.class)
-    @Context
-    public DigitalOutput createDigitalOutput(DigitalOutputConfiguration config, com.pi4j.context.Context pi4jContext) {
+    public DigitalOutput createDigitalOutput(DigitalOutputConfiguration config, Context pi4jContext) {
         var outputConfigBuilder = DigitalOutput.newConfigBuilder(pi4jContext)
                 .id(config.getId())
                 .name(config.getName())
@@ -31,7 +30,7 @@ public class Pi4JFactory {
     }
 
     @EachBean(DigitalInputConfiguration.class)
-    public DigitalInput createDigitalInput(DigitalInputConfiguration config, com.pi4j.context.Context pi4jContext) {
+    public DigitalInput createDigitalInput(DigitalInputConfiguration config, Context pi4jContext) {
         var inputConfigBuilder = DigitalInput.newConfigBuilder(pi4jContext)
                 .id(config.getId())
                 .name(config.getName())
